Use cache backend to store tqdm process information
Remove the `TaskProgressModel` model that was used to store process information. But this has some
disadvantages:

* Every `process_info.update()` call creates one `TaskProgressModel` instance and this model was
never cleaned. So if many items was processed, the table may be get full. So this PR fixed #46
* If many small `process_info.update()` calls happens, then we get a high database load

Refactor that completely by using the Django cache to store progress information.
If a task has been finished: Transfer the information from cache to database.